% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_claude.R
\docType{data}
\name{chicago}
\alias{chicago}
\title{Chicago census tracts with income data}
\format{
An \code{sf} object with 856 rows and 4 variables:
  \describe{
    \item{GEOID}{Census tract identifier}
    \item{NAME}{Census tract name}
    \item{median_income}{Median household income (USD)}
    \item{geom}{Multipolygon geometry (WGS 84)}
  }
}
\source{
Derived from US Census / ACS data.
}
\usage{
chicago
}
\description{
An \code{sf} object of census tracts for Chicago with median household
income attributes.
}
\keyword{datasets}
